third_party/blink/web_tests/external/wpt/html/infrastructure/safe-passing-of-structured-data/shared-array-buffers/window-domain-failure.https.sub-expected.txt

This is a testharness.js-based test.
FAIL SharedArrayBuffer and a same-origin-domain (but not same-origin) iframe assert_equals: expected "www1.web-platform.test" but got "web-platform.test"
Harness: the test ran to completion.
